requirements.txt
aiofiles==23.1.0
anyio==3.6.2
cmake==3.26.3
fastapi==0.95.1
ffmpeg-python==0.2.0
gunicorn==20.1.0
huggingface-hub==0.14.1
ipywidgets==8.0.6
Jinja2==3.1.2
numpy==1.24.3
openai-whisper==20230314
pydantic==1.10.7
python-magic==0.4.27
python-multipart==0.0.6
sentencepiece==0.1.99
setuptools-rust==1.6.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
uvicorn==0.22.0
